refactor: improve Python metrics API organization and cleanup
Move RuntimeMetrics implementation to prometheus_metrics.rs for better code organization. Update callback registration to use endpoint.metrics.register_update_callback(). Remove debug print statements from metrics subsystem (Python bindings, distributed runtime, and system status server).
