+++ lib/bindings/python/examples/metrics/server_with_callback.py
@@ -78,8 +78,8 @@ async def init(runtime: DistributedRuntime):
         gpu_cache_usage_perc.set(0.01 + (count * 0.01))
         print(f"[python] Updated metrics (call #{count})")
 
-    endpoint.register_metrics_callback(update_metrics)
-    print("[python] Metrics callback registered!")
+    endpoint.metrics.register_update_callback(update_metrics)
+    print("[python] update (metrics) callback registered!")
 
     # Step 3: Set initial values and test vector metrics
     print("[python] Setting initial metric values...")
